Move `ParseSink` to its own header
We will soon add a new implemenation so the one for NARs in `archive.cc`
isn't the only one.
