+++ src/libutil/archive.hh
@@ -3,6 +3,7 @@
 
 #include "types.hh"
 #include "serialise.hh"
+#include "fs-sink.hh"
 
 
 namespace nix {
@@ -71,22 +72,6 @@ time_t dumpPathAndGetMtime(const Path & path, Sink & sink,
  * @param s Contents of the file.
  */
 void dumpString(std::string_view s, Sink & sink);
-
-/**
- * \todo Fix this API, it sucks.
- */
-struct ParseSink
-{
-    virtual void createDirectory(const Path & path) { };
-
-    virtual void createRegularFile(const Path & path) { };
-    virtual void closeRegularFile() { };
-    virtual void isExecutable() { };
-    virtual void preallocateContents(uint64_t size) { };
-    virtual void receiveContents(std::string_view data) { };
-
-    virtual void createSymlink(const Path & path, const std::string & target) { };
-};
 
 /**
  * If the NAR archive contains a single file at top-level, then save
